新增log, stack, 单元测试

+++ base/fast.h
@@ -14,7 +14,7 @@ int u64toh(uint64 v, char* buf);
 
 // convert integer to ASCII string
 int u32toa(uint32 v, char* buf);
-int u64toa(uint32 v, char* buf);
+int u64toa(uint64 v, char* buf);
 
 inline int i32toa(int32 v, char* buf) {
     if (v >= 0) return u32toa((uint32)v, buf);
